requirements.txt
jquants-api-client
requests
